#529 Refactoring to cover unit tests correctly
https://github.com/metasfresh/metasfresh/issues/529
