+++ de.metas.swat/de.metas.swat.base/src/main/java/org/adempiere/invoice/service/impl/InvoiceDAO.java
@@ -30,13 +30,10 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Properties;
 
-import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.ad.trx.api.ITrx;
-import org.adempiere.invoice.service.IInvoiceBL;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.Check;
 import org.adempiere.util.LegacyAdapters;
-import org.adempiere.util.Services;
 import org.compiere.model.I_AD_Org;
 import org.compiere.model.I_C_InvoiceTax;
 import org.compiere.model.I_C_LandedCost;
@@ -201,62 +198,5 @@ public class InvoiceDAO extends AbstractInvoiceDAO
 		return result;
 	}
 
-	@Override
-	public Iterator<I_C_Invoice> retrieveCreditMemosForInvoice(final I_C_Invoice invoice)
-	{
-		List<I_C_Invoice> acreditMemos = new ArrayList<>();
 
-		final IInvoiceBL invoiceBL = Services.get(IInvoiceBL.class);
-
-		final Iterator<I_C_Invoice> referencesForInvoice = retrieveReferencesForInvoice(invoice);
-
-		while (referencesForInvoice.hasNext())
-		{
-			final I_C_Invoice referencedInvoice = referencesForInvoice.next();
-
-			if (invoiceBL.isCreditMemo(referencedInvoice))
-			{
-				acreditMemos.add(referencedInvoice);
-			}
-		}
-
-		return acreditMemos.iterator();
-	}
-
-	@Override
-	public Iterator<I_C_Invoice> retrieveAdjustmentChargesForInvoice(final I_C_Invoice invoice)
-	{
-		List<I_C_Invoice> adjustmentCharges = new ArrayList<>();
-
-		final IInvoiceBL invoiceBL = Services.get(IInvoiceBL.class);
-
-		final Iterator<I_C_Invoice> referencesForInvoice = retrieveReferencesForInvoice(invoice);
-
-		while (referencesForInvoice.hasNext())
-		{
-			final I_C_Invoice referencedInvoice = referencesForInvoice.next();
-
-			if (invoiceBL.isAdjustmentCharge(referencedInvoice))
-			{
-				adjustmentCharges.add(referencedInvoice);
-			}
-		}
-
-		return adjustmentCharges.iterator();
-	}
-
-	
-	private Iterator<I_C_Invoice> retrieveReferencesForInvoice(final I_C_Invoice invoice)
-	{
-		// services
-		final IQueryBL queryBL = Services.get(IQueryBL.class);
-
-		return queryBL.createQueryBuilder(I_C_Invoice.class, invoice)
-				.filterByClientId()
-				.addOnlyActiveRecordsFilter()
-				.addEqualsFilter(I_C_Invoice.COLUMNNAME_Ref_Invoice_ID, invoice.getC_Invoice_ID())
-
-				.create()
-				.iterate(I_C_Invoice.class);
-	}
 }
