+++ de.metas.swat/de.metas.swat.base/src/main/java/org/adempiere/invoice/service/impl/PlainInvoiceDAO.java
@@ -241,16 +241,4 @@ public class PlainInvoiceDAO extends AbstractInvoiceDAO
 
 		});
 	}
-
-	@Override
-	public Iterator<I_C_Invoice> retrieveAdjustmentChargesForInvoice(I_C_Invoice invoice)
-	{
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Iterator<I_C_Invoice> retrieveCreditMemosForInvoice(I_C_Invoice invoice)
-	{
-		throw new UnsupportedOperationException();
-	}
 }
